Add user helper, update user model and change user function
Added a user helper function that retrieves the current user data based on the JWT from the header.

Changed JWT token to JWT in an error string to remove redundancy.

Added a GetUserByID function that retrieves user data for a provided ID.

Changed QueryUsername to GetUserByName for clarity and consistency.