Add view message and update message structure
Added the view message function and incorporated it into the protected API route.

Added the GetMessageByID function.

Moved the CORS middleware to an independent file in the middleware folder.

Changed uid to id for the GetUserByID function for clarity.

Updated the message structure to include message subject, and set the message string to not null.