Update store_test.go
Updated the retrieval of the JWT for unit testing. The test now decodes the response body. This was necessary because the login API was changed to provide the front end with the username and userid for caching.

Refactored the cors middleware to include an access control URL and to conform to code standards.