Refactor user messaging
Changed the model structure of messaging to include a one-to-many association for responses. Included a one-to-one association for sender and receiver.

Amended the user model to include a one-to-many association for sent messages and received messages.

Sender, receiver and replies are preloaded when utilizing GetMessageByID.

Routing for sendmessage and viewmessage have been added to the protectedAPI group.