Add transaction refactor product model
Added a transaction model for tracking sales history.

Added table auto migration to main.go for transactions.

Updated the product model to include quantity. Changed originalPrice to price and removed finalPrice. The finalPrice field now exists in transactions.

Added a one-to-many association with transactions for purchases and sales to the user model.