Add user inbox and sent messages
Added the protected API routes for user inbox and user sent messages.

Added GetUserInbox to the user model.

Added GetUserSentMessages to the user model.

Removed the one-to-many association for the user inbox and sent messages as this functionality is no longer needed.

Updated product search to sort by descending order.